# docs/swagger.yaml
basePath: /
definitions:
...
  models.Branches:
    properties:
      branch_id:
        type: integer
      branch_name:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Comments:
    properties:
      branch:
        allOf:
        - $ref: '#/definitions/models.Branches'
        description: 关联关系
      branch_id:
        type: integer
      comment_content:
        type: string
      comment_id:
        type: integer
      course_id:
        type: integer
      created_at:
        type: string
      parent_comment:
        $ref: '#/definitions/models.Comments'
      parent_comment_id:
        description: 回复的评论ID
        type: integer
      replies:
        items:
          $ref: '#/definitions/models.Comments'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
...
  models.Learning:
    properties:
      completed_at:
        type: string
      course_id:
        type: integer
      created_at:
        type: string
      learning_id:
        type: integer
      progress_percentage:
        type: integer
      status:
        description: enrolled, in_progress, completed
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.Users'
        description: 关联关系
      user_id:
        type: integer
    type: object
...
  models.Users:
    properties:
      branch:
        allOf:
        - $ref: '#/definitions/models.Branches'
        description: 关联关系
      branch_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      role:
        description: student, teacher
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  service.AddCommentRequest:
    properties:
      content:
        type: string
      parent_comment_id:
        type: integer
    required:
    - content
    type: object
...
  service.CommentView:
    properties:
      branch_id:
        type: integer
      comment_content:
        type: string
      comment_id:
        type: integer
      course_id:
        type: integer
      created_at:
        type: string
      parent_comment_id:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  service.LearningProgressView:
    properties:
      branch_id:
        type: integer
      completed_at:
        type: string
      course_id:
        type: integer
      email:
        type: string
      progress_percentage:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  service.UpdateProgressRequest:
    properties:
      progress_percentage:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 这是一个分布式在线学习平台的API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 分布式在线学习平台 API
  version: "1.0"
paths:
  /api/v1/courses/{id}/comments:
    get:
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.CommentView'
            type: array
      summary: 获取课程评论
      tags:
      - 评论
...
  /api/v1/student/courses/{id}/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddCommentRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comments'
      security:
      - BearerAuth: []
      summary: 学生发表评论
      tags:
      - 学生评论
  /api/v1/student/courses/{id}/enroll:
    post:
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Learning'
      security:
      - BearerAuth: []
      summary: 学生报名课程
      tags:
      - 学生学习
  /api/v1/student/courses/{id}/progress:
    get:
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Learning'
      security:
      - BearerAuth: []
      summary: 获取学习进度
      tags:
      - 学生学习
    put:
      consumes:
      - application/json
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      - description: 学习进度
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateProgressRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Learning'
      security:
      - BearerAuth: []
      summary: 更新学习进度
      tags:
      - 学生学习
...
  /api/v1/teacher/courses/{id}/learning:
    get:
      parameters:
      - description: 课程ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.LearningProgressView'
            type: array
      security:
      - BearerAuth: []
      summary: 查看课程学习进度
      tags:
      - 教师学习
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"